[COS-578] - implemented and moved the wait mechanisms
