+++ packages/cli/src/commands/dev/start.ts
@@ -49,7 +49,6 @@ export default class DevStart extends Command {
         'file-store-api',
         'jaeger',
         'settings-api',
-        'test-contract',
         'test-env',
         `user-admin-api`,
         'validation-api',
@@ -94,6 +93,7 @@ export default class DevStart extends Command {
       installJaeger(flags.verbose, location, version)
       sql.provisionPostgresql(flags.verbose, location)
       neo.provisionNeo4j(flags.verbose, location)
+      neo.provisionNeo4jWithDemoTenant(flags.verbose)
       redis.provisionRedis(flags.verbose, location)
       start.cleanupSetupFiles()
       start.cleanupSetupFiles()
@@ -123,6 +123,7 @@ export default class DevStart extends Command {
       installJaeger(flags.verbose, location, version)
       sql.provisionPostgresql(flags.verbose, location)
       neo.provisionNeo4j(flags.verbose, location)
+      neo.provisionNeo4jWithDemoTenant(flags.verbose)
       redis.provisionRedis(flags.verbose, location)
       start.cleanupSetupFiles()
       logTerminal('INFO', 'to ensure everything was installed correctly, run => openline dev ping')
@@ -176,6 +177,7 @@ export default class DevStart extends Command {
       start.installDatabases(flags.verbose, location)
       sql.provisionPostgresql(flags.verbose, location)
       neo.provisionNeo4j(flags.verbose, location)
+      neo.provisionNeo4jWithDemoTenant(flags.verbose)
       redis.provisionRedis(flags.verbose, location)
       start.cleanupSetupFiles()
       logTerminal('INFO', 'to ensure everything was installed correctly, run => openline dev ping')
@@ -246,21 +248,7 @@ export default class DevStart extends Command {
         installUserAdminApi(flags.verbose, location, version)
         sql.provisionPostgresql(flags.verbose, location)
         neo.provisionNeo4j(flags.verbose, location)
-        redis.provisionRedis(flags.verbose, location)
-        start.cleanupSetupFiles()
-        logTerminal('INFO', 'to ensure everything was installed correctly, run => openline dev ping')
-        break
-
-      case 'test-contract':
-        start.dependencyCheck(flags.verbose)
-        start.startDevServer(flags.verbose)
-        start.cleanupSetupFiles()
-        ns.installNamespace(flags.verbose, location)
-        start.installDatabases(flags.verbose, location)
-        installCustomerOsApi(flags.verbose, location, version)
-        installUserAdminApi(flags.verbose, location, version)
-        sql.provisionPostgresql(flags.verbose, location)
-        neo.provisionNeo4j(flags.verbose, location)
+        neo.provisionNeo4jWithDemoTenant(flags.verbose)
         redis.provisionRedis(flags.verbose, location)
         start.cleanupSetupFiles()
         logTerminal('INFO', 'to ensure everything was installed correctly, run => openline dev ping')
