+++ packages/cli/src/lib/dev/customer-os.ts
@@ -3,7 +3,6 @@ import {getConfig} from '../../config/dev'
 import {deployImage, Yaml, updateImageTag} from './deploy'
 import {buildLocalImage} from './build-image'
 import {logTerminal} from '../logs'
-import {waitForNeo4jToBeInitialized} from "./neo4j";
 
 const config = getConfig()
 const NAMESPACE = config.namespace.name
@@ -257,46 +256,6 @@ export function installValidationApi(verbose: boolean, location = config.setupDi
   return true
 }
 
-function provisionNeo4jWithDemoTenant() {
-  logTerminal('INFO', 'Waiting for Neo4J to be provisioned with the demo tenant')
-  shell.exec('sleep 2')
-  const axios = require('axios');
-  const FormData = require('form-data');
-  const fs = require('fs');
-
-  const url = 'http://127.0.0.1:4001/demo-tenant';
-  const headers = {
-    'X-Openline-Api-Key': 'cad7ccb6-d8ff-4bae-a048-a42db33a217e',
-    'TENANT_NAME': 'openline',
-    'MASTER_USERNAME': 'development@openline.ai',
-  };
-
-  const form = new FormData();
-
-// Fetch the JSON data from the URL
-  axios.get('https://raw.githubusercontent.com/openline-ai/openline-cli/otter/resources/demo-tenant.json')
-    .then((response: import('axios').AxiosResponse) => {
-      // Append the fetched data to the form
-      form.append('file', Buffer.from(JSON.stringify(response.data)), {
-        filename: 'demo-tenant.json',
-        contentType: 'application/json',
-      });
-      return axios({
-        method: 'get',
-        url,
-        headers,
-        data: form,
-        maxRedirects: 0,
-      });
-    })
-    .then((response: import('axios').AxiosResponse) => {
-      console.log('Response:', response.data);
-    })
-    .catch((error: Error) => {
-      console.error('Error:', error.message);
-    });
-}
-
 export function installUserAdminApi(verbose: boolean, location = config.setupDir, imageVersion = 'latest') :boolean {
   if (userAdminApiCheck()) {
     logTerminal('SUCCESS', 'user-admin-api already running')
@@ -335,15 +294,11 @@ export function installUserAdminApi(verbose: boolean, location = config.setupDir
   const deploy = deployImage(image, installConfig, verbose)
   if (deploy === false) return false
 
-  waitForUserAdminAppPodToBeReady();
-  waitForNeo4jToBeInitialized(verbose)
-  provisionNeo4jWithDemoTenant();
-
   logTerminal('SUCCESS', 'user-admin-api successfully installed')
   return true
 }
 
-function waitForUserAdminAppPodToBeReady() {
+export function waitForUserAdminAppPodToBeReady() {
   logTerminal('INFO', 'Waiting for user-admin-api pod to be Ready')
   let userAdminApiPodName
   do {
